@@ -1,58 +1,51 @@
-var tape = require("tape"),
-    d3 = require("../"),
-    d3_geo = require("d3-geo");
-require("./projectionEqual");
+import * as d3 from "../src/index.js";
+import {assertProjectionEqual} from "./asserts.js";
 
-
-tape("inverse polyhedrals", function(test) {
+it("inverse polyhedrals", () => {
 
   [ d3.geoAirocean(), d3.geoCubic(), d3.geoIcosahedral(), d3.geoDodecahedral() ]
   .forEach(function(projection) {
     [ [-23, 12], [10,10], [100,-45] ]
     .forEach(function(location) {
       projection.angle(Math.random()*360);
-      test.projectionEqual(projection, location, projection(location), 1e-5);
+      assertProjectionEqual(projection, location, projection(location), 1e-5);
     });
   });
-  test.end();
 });
 
-tape("inverse Imago, tetrahedralLee", function(test) {
+it("inverse Imago, tetrahedralLee", () => {
 
   [ d3.geoImago(), d3.geoImagoBlock(), d3.geoTetrahedralLee() ]
   .forEach(function(projection) {
     [ [-23, 12], [10,10], [100,-45] ]
     .forEach(function(location) {
       projection.angle(Math.random()*360);
-      test.projectionEqual(projection, location, projection(location), 1e-5);
+      assertProjectionEqual(projection, location, projection(location), 1e-5);
     });
   });
-  test.end();
 });
 
-tape("inverse Cahill-Keyes", function(test) {
+it("inverse Cahill-Keyes", () => {
 
   [ d3.geoCahillKeyes() ]
   .forEach(function(projection) {
     [ [-23, 12], [10,10], [100,-45] ]
     .forEach(function(location) {
       projection.angle(Math.random()*360);
-      test.projectionEqual(projection, location, projection(location), 1e-5);
+      assertProjectionEqual(projection, location, projection(location), 1e-5);
     });
   });
-  test.end();
 });
 
-tape("inverse complex log", function(test) {
+it("inverse complex log", () => {
 
   [ d3.geoComplexLog() ]
   .forEach(function(projection) {
     [ [0, 0], [-23, 12], [10,10], [100,-45] ]
     .forEach(function(location) {
       projection.angle(Math.random()*360);
-      test.projectionEqual(projection, location, projection(location), 1e-5);
+      assertProjectionEqual(projection, location, projection(location), 1e-5);
     });
   });
-  test.end();
 });
 
