@@ -1,17 +1,17 @@
-export {default as geoClipPolygon} from "./clip/polygon";
-export {default as geoIntersectArc} from "./intersect";
-export {default as geoPolyhedral} from "./polyhedral/index";
-export {default as geoPolyhedralButterfly} from "./polyhedral/butterfly";
-export {default as geoPolyhedralCollignon} from "./polyhedral/collignon";
-export {default as geoPolyhedralWaterman} from "./polyhedral/waterman";
-export {default as geoPolyhedralVoronoi} from "./polyhedral/voronoi";
-export {default as geoDodecahedral} from "./polyhedral/dodecahedral";
+export {default as geoClipPolygon} from "./clip/polygon.js";
+export {default as geoIntersectArc} from "./intersect.js";
+export {default as geoPolyhedral} from "./polyhedral/index.js";
+export {default as geoPolyhedralButterfly} from "./polyhedral/butterfly.js";
+export {default as geoPolyhedralCollignon} from "./polyhedral/collignon.js";
+export {default as geoPolyhedralWaterman} from "./polyhedral/waterman.js";
+export {default as geoPolyhedralVoronoi} from "./polyhedral/voronoi.js";
+export {default as geoDodecahedral} from "./polyhedral/dodecahedral.js";
 export {default as geoCox, coxRaw as geoCoxRaw} from "./cox.js";
 export {default as geoTetrahedralLee, leeRaw as geoLeeRaw} from "./tetrahedralLee.js";
-export {default as geoGrayFullerRaw} from "./grayfuller";
-export {default as geoAirocean} from "./airocean";
-export {default as geoIcosahedral} from "./icosahedral";
-export {default as geoImago, imagoBlock as geoImagoBlock, imagoRaw as geoImagoRaw} from "./imago";
-export {default as geoCubic} from "./cubic";
-export {default as geoCahillKeyes, cahillKeyesRaw as geoCahillKeyesRaw} from "./cahillKeyes";
-export {default as geoComplexLog, complexLogRaw as geoComplexLogRaw} from "./complexLog";
+export {default as geoGrayFullerRaw} from "./grayfuller.js";
+export {default as geoAirocean} from "./airocean.js";
+export {default as geoIcosahedral} from "./icosahedral.js";
+export {default as geoImago, imagoBlock as geoImagoBlock, imagoRaw as geoImagoRaw} from "./imago.js";
+export {default as geoCubic} from "./cubic.js";
+export {default as geoCahillKeyes, cahillKeyesRaw as geoCahillKeyesRaw} from "./cahillKeyes.js";
+export {default as geoComplexLog, complexLogRaw as geoComplexLogRaw} from "./complexLog.js";
